backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
python-jose
python-multipart
requests